move some fabric stuff
