+++ loadercomplex-fabric-17/src/main/java/com/enderzombi102/loadercomplex/fabric17/impl/block/FabricBlock.java
@@ -4,7 +4,7 @@ import com.enderzombi102.loadercomplex.api.block.Block;
 import com.enderzombi102.loadercomplex.fabric17.impl.entity.FabricEntity;
 import com.enderzombi102.loadercomplex.fabric17.impl.entity.FabricPlayer;
 import com.enderzombi102.loadercomplex.fabric17.impl.utils.BlockUtils;
-import com.enderzombi102.loadercomplex.fabric17.impl.utils.Callable;
+import com.enderzombi102.loadercomplex.Callable;
 import com.enderzombi102.loadercomplex.fabric17.impl.world.FabricWorld;
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.BlockState;
