+++ loadercomplex-fabric-common/src/main/resources/assets/loadercomplex/icon4fabric.png
